Expand intersection offsets to 64-bit (#708)
The number of intersections is proportional to the product of the image
resolution and number of Gaussians. Thus, the intersection offsets are
the largest linear index space within Gaussian splat training pipeline
and can easily overflow for a large geospatial scene with
high-resolution photography where we refine to obtain fine details. The
overflow results in the rasterization operators accessing
invalid/concurrently written memory which eventually leads to NaNs in
the loss (although the Gaussian parameters blow up long before that.)

This PR addresses this by switching to 64-bit intersection offsets and
indices and fixes the incorrect training results.

---------

+++ src/tests/GaussianRasterizeContributingGaussianIdsTest.cpp
@@ -31,7 +31,7 @@ struct GaussianRasterizeContributingGaussianIdsTestFixture : public ::testing::T
         means2d                           = inputs[0].cuda();
         conics                            = inputs[1].cuda();
         opacities                         = inputs[2].cuda();
-        tileOffsets                       = inputs[3].cuda();
+        tileOffsets                       = inputs[3].to(torch::kInt64).cuda();
         tileGaussianIds                   = inputs[4].cuda();
         imageDims                         = inputs[5];
